valid argument, movies in a single file
